refactored and added death stats.
