+++ package.json
@@ -1,20 +1,20 @@
 {
-  "name": "flappy-tinybird",
-  "description": "This repository contains a clone of the popular game 'Flappy Bird'. It was built using the Phaser 3 game framework and JavaScript.",
-  "private": false,
-  "version": "0.0.1",
-  "type": "module",
-  "scripts": {
-    "start": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "phaser": "^3.60.0",
-    "uuid": "^9.0.0"
-  },
-  "devDependencies": {
-    "eslint": "^8.41.0",
-    "vite": "^4.3.8"
-  }
+    "name": "flappy-tinybird",
+    "description": "This repository contains a clone of the popular game 'Flappy Bird'. It was built using the Phaser 3 game framework and JavaScript.",
+    "private": false,
+    "version": "0.0.1",
+    "type": "module",
+    "scripts": {
+        "start": "vite",
+        "build": "vite build",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "phaser": "^3.60.0",
+        "uuid": "^9.0.0"
+    },
+    "devDependencies": {
+        "eslint": "^8.41.0",
+        "vite": "^4.3.8"
+    }
 }
